refactor(create): move textinput to questionList component
